requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
altair>=5.0.0
openpyxl>=3.1.0
scipy>=1.10.0
plotly>=5.0.0
python-dotenv>=1.0.0
openai>=1.0.0
